refactor(runtime): split reaper.rs god object into modules
