+++ scripts/.god-object-baseline.txt
@@ -6,7 +6,6 @@
 1065 sandbox-runtime/src/session_auth.rs
 948 ai-agent-instance-blueprint-lib/src/workflows.rs
 922 ui/src/routes/workflows._index.tsx
-838 sandbox-runtime/src/reaper.rs
 744 sandbox-runtime/src/scoped_session_auth.rs
 738 ui/src/routes/sandboxes.$id.tsx
 721 ui/src/components/shared/SessionSidebar.tsx
